imporve 1

+++ pages/listing.php
@@ -1,26 +1,17 @@
-<?php 	
+<?php
 	if (isset($_GET['id'])){
-		$result = $mysqli->query("select image from `posting` where id=" . $_GET['id']);
+		$result = $mysqli->query("select url from `images` where postid=" . $_GET['id']);
 		if ($result){
 			$row = $result->fetch_assoc();
 		}
 		if (isset($row)){
-			$image = $row['image'];
-			if (file_exists($image))
-							unlink($image);
+			$image = $row['url'];
+			if (file_exists($image)) unlink($image);
 		}
-		$query = "delete from `posting` where id=" . $_GET['id'];
-		$mysqli->query($query);
+		$mysqli->query("delete from `posting` where id=" . $_GET['id']);
+		$mysqli->query("delete from `images` where postid=" . $_GET['id']);
 	}
-
-	$query = "SELECT * FROM `posting` ORDER BY `timestamps` DESC";
-	$result = $mysqli->query($query);
-	if ($result){
-		$data = [];
-		while($row = $result->fetch_assoc()){
-			array_push($data, $row);
-		}
-	}
+	include "components/post_list.php";
 ?>
 
 
@@ -32,19 +23,25 @@ if (count($data) > 0){
 		<p>
 			<i class="far fa-clock"></i>
 			<?php echo $row['timestamps']; ?>
+			<span style="float:right">
+				By
+				<strong>
+					<?php echo $row['username'];?>
+				</strong>
+			</span>
 		</p>
 	</div>
 	<div class="panel-body">
-		<?php if ($row['image'] != "") {?>
-		<div class="col-md-5">
-			<img src="<?php  echo $row['image']?>" style="max-width: 100%" alt="<?php  echo $row['image']?>">
-		</div>
-		<div class="col-md-7">
-			<?php echo nl2br($row['content']); ?>
-		</div>
-		<?php } else echo nl2br($row['content']); ?>
-		<div class="col-md-12">
-			<a href="?page=listing&id=<?php echo $row['id'];?>" class="delete">Delete</a>
+		<?php if (count($row['images']) > 0) {
+			echo '<div class="mosaic">';
+			foreach ($row['images'] as $image){
+			?>
+		<img src="<?php  echo $image?>" alt="<?php  echo $image?>">
+		<?php }
+		echo '</div><div class="col-md-12 row"><hr>' . nl2br($row['content']).'</div>';
+	} else echo nl2br($row['content']); ?>
+		<div class="col-md-12 row">
+			<a href="?page=listing&id=<?php echo $row['postid'];?>" class="delete">Delete</a>
 		</div>
 	</div>
 </div>
